feat: 읽기 페이지 전용 헤더 구현
